requirements.txt
scrapybara
undetected_playwright_patch
